Adding solution to Median of the two array

+++ MedianofTwoArray-29032022/Solution.java
@@ -1,8 +1,8 @@
 public class Solution {
     public double findMedianSortedArrays(int[] nums1, int[] nums2) {
-        double size1=nums1.length;
-        double size2=nums2.length;
-        double size=size1+size2;
+        int  size1=nums1.length;
+        int size2=nums2.length;
+        int size=size1+size2;
         double median=0;
         if(size%2==0){
             median=nums1[size/2];
